requirements.txt
requests
python-dotenv
psycopg2-binary

langchain-community
langchain-openai